+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - New: Watchlist, auto corporation whitelist.
 - Change: Trim character search input.
+- Integrated Swagger UI installation with frontend.
 - Other small stuff.
 
 ## 1.8.0